+++ sdk/identity/azure-identity-broker/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Features Added
 
+- Added support for using the default broker 
+
 ### Breaking Changes
 
 ### Bugs Fixed